+++ src/index.css
@@ -15,7 +15,7 @@ body{
 html {
     scroll-behavior: smooth;
 }
-.learnMoreBtn__container{
+.Btn__container{
     display: flex;
     align-items: center;
     justify-content: space-around;
